requirements.txt
backports.functools-lru-cache==1.5
certifi==2018.10.15
chardet==3.0.4
cntk==2.6
cycler==0.10.0
Django==1.11.16
djangorestframework==3.8.2
enum34==1.1.6
h5py==2.8.0
idna==2.7
jieba==0.39
Keras==2.2.4
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
matplotlib==2.2.3
MySQL-python==1.2.5
nltk==3.3
numpy==1.15.2
pandas==0.23.4
pyparsing==2.3.0
python-dateutil==2.7.3
pytz==2018.5
PyYAML==3.13
requests==2.19.1
scikit-learn==0.20.0
scipy==1.1.0
six==1.11.0
snownlp==0.12.3
Theano==1.0.3
urllib3==1.23
zhon==1.1.5
